增加构建 rootfs 镜像

+++ scripts/build-image.sh
@@ -13,9 +13,7 @@ RESERVED2_SIZE=8192
 LOADER2_SIZE=8192
 TRUST_SIZE=8192
 BOOT_SIZE=229376
-#BOOT_SIZE=1048576
 
-#SYSTEM_START=0
 LOADER1_START=64                                      # 64 preloader (miniloader or U-Boot SPL)
 RESERVED1_START=$((LOADER1_START + LOADER1_SIZE))     # 8064 legacy DRM key
 RESERVED2_START=$((RESERVED1_START + RESERVED1_SIZE)) # 8192 legacy parameter
@@ -23,7 +21,6 @@ LOADER2_START=$((RESERVED2_START + RESERVED2_SIZE))   # 16384 U-Boot or UEFI
 TRUST_START=$((LOADER2_START + LOADER2_SIZE))         # 24576 trusted-os like ATF, OP-TEE
 BOOT_START=$((TRUST_START + TRUST_SIZE))              # 32768 kernel, dtb, extlinux.conf, ramdisk
 ROOTFS_START=$((BOOT_START + BOOT_SIZE))              # 262144 linux system
-#ROOTFS_START=$((BOOT_START + BOOT_SIZE))              # 1081344 linux system
 
 SCRIPTS_PATH=$(
     cd "$(dirname "$0")"
@@ -31,19 +28,20 @@ SCRIPTS_PATH=$(
 )
 PROJECT_PATH=$(dirname "$SCRIPTS_PATH")
 OUTPUT_PATH=$PROJECT_PATH/out
+PWD_PATH=$(pwd)
 
 TARGET=$1
+ROOTFS_PATH=$PWD_PATH/$2
 
 # 构建完整镜像所必须的文件
-#$OUTPUT_PATH
-#├── kernel
-#│   ├── Image
-#│   └── tn3399-linux.dtb
-#├── rootfs.img
-#└── u-boot
-#    ├── idbloader.img
-#    ├── trust.img
-#    └── u-boot.img
+# $OUTPUT_PATH
+# ├── kernel
+# │   ├── Image
+# │   └── tn3399-linux.dtb
+# └── u-boot
+#     ├── idbloader.img
+#     ├── trust.img
+#     └── u-boot.img
 
 # u-boot
 IDBLOADER_IMGAGE=$OUTPUT_PATH/u-boot/idbloader.img
@@ -64,6 +62,12 @@ build_trust() {
 
 build_boot() {
     rm -f "$BOOT_IMAGE"
+
+    if [ ! -s "$BOOT_IMAGE" ]; then
+        echo "BUILD FAILED: MISSING BOOT FILES"
+        exit 1
+    fi
+
     cat >"$OUTPUT_PATH"/tn3399.conf <<EOF
 label kernel-5.4
     kernel /Image
@@ -86,10 +90,56 @@ EOF
     # 将 kernel 编译出的 Image 和 dtb 文件复制到分区
     mcopy -i "$BOOT_IMAGE" -s "$DTB_FILE" ::tn3399-linux.dtb
     mcopy -i "$BOOT_IMAGE" -s "$KERNEL_IMAGE" ::Image
+
+    echo "BUILD SUCCEED: $BOOT_IMAGE"
+}
+
+build_rootfs() {
+    rm -f "$ROOTFS_IMAGE"
+
+    if [ ! -d "$ROOTFS_PATH" ]; then
+        echo "BUILD FAILED: ROOTFS PATH NOT SET"
+        exit 1
+    fi
+
+    # 创建 2G 空白文件
+    dd if=/dev/zero of="$ROOTFS_IMAGE" bs=1M count=2048 oflag=sync status=progress
+    # 格式化为 ext4 分区
+    mkfs.ext4 "$ROOTFS_IMAGE"
+
+    tmp=$(mktemp -d)
+    sudo mount "$ROOTFS_IMAGE" "$tmp"
+    # 复制 rootfs 文件到镜像中
+    sudo cp -rfp "$ROOTFS_PATH"/* "$tmp"
+    sudo umount "$tmp"
+    rm -r "$tmp"
+
+    # 检查并修复文件系统
+    e2fsck -p -f "$ROOTFS_IMAGE"
+    # 压缩镜像
+    resize2fs -M "$ROOTFS_IMAGE"
+
+    echo "BUILD SUCCEED: $ROOTFS_IMAGE"
 }
 
 build_system() {
     rm -f "$SYSTEM_IMAGE"
+
+    if [ ! -s "$IDBLOADER_IMGAGE" ] || [ ! -s "$UBOOT_IMGAGE" ] && [ ! -s "$TRUSRT_IMGAGE" ]; then
+        echo "BUILD FAILED: MISSING U-BOOT IMAGE"
+        exit 1
+    fi
+
+    if [ ! -s "$BOOT_IMAGE" ]; then
+        echo "BUILD FAILED: MISSING BOOT IMAGE"
+        exit 1
+    fi
+
+    if [ ! -s "$ROOTFS_IMAGE" ]; then
+        echo "BUILD FAILED: MISSING ROOTFS IMAGE"
+        exit 1
+    fi
+
     # 计算镜像大小
     ROOTFS_SIZE=$(stat -L --format="%s" "$ROOTFS_IMAGE")
     GPT_IMAGE_SIZE_MIN=$(((LOADER1_SIZE + RESERVED1_SIZE + RESERVED2_SIZE + LOADER2_SIZE + TRUST_SIZE + BOOT_SIZE + 35) * 512 + ROOTFS_SIZE))
@@ -130,12 +180,22 @@ EOF
 
     # 烧写 rootfs
     dd if="$ROOTFS_IMAGE" of="$SYSTEM_IMAGE" seek=$ROOTFS_START conv=notrunc,fsync
+
+    echo "BUILD SUCCEED: $SYSTEM_IMAGE"
 }
 
 if [ "$TARGET" == "trust" ]; then
     build_trust
 elif [ "$TARGET" == "boot" ]; then
     build_boot
+elif [ "$TARGET" == "rootfs" ]; then
+    build_rootfs
 elif [ "$TARGET" == "system" ]; then
     build_system
+else
+    echo
+    echo "usage:"
+    echo "build_image.sh <trust | boot | system>"
+    echo "build_image.sh <rootfs> <rootfs-relative-path>"
+    echo
 fi
